+++ CHANGELOG.md
@@ -61,6 +61,7 @@ You can find its changes [documented below](#070---2021-01-01).
 - `scroll_to_view` and `scroll_area_to_view` methods on `UpdateCtx`, `LifecycleCtx` and `EventCtx` ([#1976] by [@xarvic])
 - `Notification::route` ([#1978] by [@xarvic])
 - Build on OpenBSD ([#1993] by [@klemensn])
+- `Event::WindowFocus(bool)` to notify widgets of window focus changes ([#2016] by [@ForLoveOfCats])
 
 ### Changed
 
@@ -804,6 +805,7 @@ Last release without a changelog :(
 [#1978]: https://github.com/linebender/druid/pull/1978
 [#1993]: https://github.com/linebender/druid/pull/1993
 [#1996]: https://github.com/linebender/druid/pull/1996
+[#2016]: https://github.com/linebender/druid/pull/2016
 
 [Unreleased]: https://github.com/linebender/druid/compare/v0.7.0...master
 [0.7.0]: https://github.com/linebender/druid/compare/v0.6.0...v0.7.0